+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: hroh <hroh@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/05 10:48:15 by hroh              #+#    #+#             */
-/*   Updated: 2020/10/06 13:56:53 by hroh             ###   ########.fr       */
+/*   Updated: 2020/10/08 16:04:12 by hroh             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *str1, const char *str2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (i < n && str1[i])
